requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
sendgrid==6.10.0
python-http-client==3.3.7
gunicorn==21.2.0
Flask-HTTPAuth==4.8.0
